banhang

+++ BanMayTinh/Migrations/DataContextModelSnapshot.cs
@@ -155,8 +155,8 @@ namespace BanMayTinh.Migrations
                     b.Property<decimal>("Price")
                         .HasColumnType("decimal(18,2)");
 
-                    b.Property<int>("ProductId")
-                        .HasColumnType("int");
+                    b.Property<long>("ProductId")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("Quantity")
                         .HasColumnType("int");
